refactor(workforce): extract reference seed data
Docs-Impact-Decision: Internal seed-module extraction preserves the workforce-seed public API and changes no user-visible behavior.
